requirements.txt
pandas
geopandas
shapely
numpy
Pillow
requests
opencv-python